client side cleaned up in client side, react app lang from js changed to ts
