+++ exp-client/src/App.tsx
@@ -1,0 +1,9 @@
+function App() {
+  return (
+    <>
+      <h1>hey</h1>
+    </>
+  );
+}
+
+export default App;
